Fix the existing CA support It seems to have broken at some point.
This is verified to fix the problem by creating a cluster with a configuration which requires ca-key.pem to be deployed on controller nodes:

```
cd e2e && EXISTING_CA=1 RBAC=1 NODE_AUTHORIZER=1 KUBE_AWS_USE_CALICO=1 ETCD_DISASTER_RECOVERY_AUTOMATED=1 ETCD_SNAPSHOT_AUTOMATED=1 ETCD_COUNT=1 CONTROLLER_COUNT=1 KUBE_AWS_CLUSTER_NAME=clustername bash -c './run all'
```

and an another one which doesn't require ca-key.pem on controller nodes:

```
cd e2e && EXISTING_CA=1 ETCD_COUNT=1 CONTROLLER_COUNT=1 KUBE_AWS_CLUSTER_NAME=clustername bash -c './run all'
```

Fixes #839

+++ test/helper/helper.go
@@ -20,6 +20,14 @@ func WithTempDir(fn func(dir string)) {
 }
 
 func WithDummyCredentials(fn func(dir string)) {
+	withDummyCredentials(true, fn)
+}
+
+func WithDummyCredentialsButCAKey(fn func(dir string)) {
+	withDummyCredentials(false, fn)
+}
+
+func withDummyCredentials(alsoWriteCAKey bool, fn func(dir string)) {
 	dir, err := ioutil.TempDir("", "dummy-credentials")
 
 	if err != nil {
@@ -38,11 +46,13 @@ func WithDummyCredentials(fn func(dir string)) {
 		}
 		defer os.Remove(certFile)
 
-		keyFile := fmt.Sprintf("%s/%s-key.pem", dir, pairName)
-		if err := ioutil.WriteFile(keyFile, []byte("dummykey"), 0644); err != nil {
-			panic(err)
+		if pairName != "ca" || alsoWriteCAKey {
+			keyFile := fmt.Sprintf("%s/%s-key.pem", dir, pairName)
+			if err := ioutil.WriteFile(keyFile, []byte("dummykey"), 0644); err != nil {
+				panic(err)
+			}
+			defer os.Remove(keyFile)
 		}
-		defer os.Remove(keyFile)
 	}
 
 	symlinks := []struct {
